requirements.txt
blinker==1.9.0
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.1.0
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
sqlalchemy 
pymysql 
cloud-sql-python-connector[pytds]
google-cloud-storage